Implementation of logs that copies from Raft to a separate file per log
This is cleaner from the standpoint of Raft, but also solves the multi-log problem,
at the expense of writing everything twice.  But it's not really writes that hurt,
it's probably flushes-to-disk.  This way all logs can share a sync into the Raft log,
and then snapshots can be built with only a single flush required.